+++ src/index.ts
@@ -2,6 +2,8 @@ import { concatBytes } from "@noble/ciphers/utils";
 
 import { type Config, ECIES_CONFIG } from "./config.js";
 import { PrivateKey, PublicKey } from "./keys/index.js";
+
+import { IS_BUFFER_SUPPORTED } from "./types.js";
 import {
   aesDecrypt,
   aesEncrypt,
@@ -14,15 +16,21 @@ import {
 
 /**
  * Encrypts data with a receiver's public key.
- * @description From version 0.5.0, `Uint8Array` will be returned instead of `Buffer`.
- * To keep the same behavior, use `Buffer.from(encrypt(...))`.
+ * @description
+ * In version 0.4.18, `Buffer` is returned when available, otherwise `Uint8Array`.
+ * From version 0.5.0, this function will always return `Uint8Array`.
+ * To preserve the pre-0.5.0 behavior of returning a `Buffer`, wrap the result with `Buffer.from(encrypt(...))`.
  *
  * @param receiverRawPK - Raw public key of the receiver, either as a hex `string` or a `Uint8Array`.
  * @param data - Data to encrypt.
  * @returns Encrypted payload, format: `public key || encrypted`.
  */
-export function encrypt(receiverRawPK: string | Uint8Array, data: Uint8Array): Buffer {
-  return Buffer.from(_encrypt(receiverRawPK, data, ECIES_CONFIG));
+export function encrypt(
+  receiverRawPK: string | Uint8Array,
+  data: Uint8Array
+): Uint8Array {
+  const encrypted = _encrypt(receiverRawPK, data, ECIES_CONFIG);
+  return IS_BUFFER_SUPPORTED ? Buffer.from(encrypted) : encrypted;
 }
 
 function _encrypt(
@@ -47,15 +55,21 @@ function _encrypt(
 
 /**
  * Decrypts data with a receiver's private key.
- * @description From version 0.5.0, `Uint8Array` will be returned instead of `Buffer`.
- * To keep the same behavior, use `Buffer.from(decrypt(...))`.
+ * @description
+ * In version 0.4.18, `Buffer` is returned when available, otherwise `Uint8Array`.
+ * From version 0.5.0, this function will always return `Uint8Array`.
+ * To preserve the pre-0.5.0 behavior of returning a `Buffer`, wrap the result with `Buffer.from(decrypt(...))`.
  *
  * @param receiverRawSK - Raw private key of the receiver, either as a hex `string` or a `Uint8Array`.
  * @param data - Data to decrypt.
  * @returns Decrypted plain text.
  */
-export function decrypt(receiverRawSK: string | Uint8Array, data: Uint8Array): Buffer {
-  return Buffer.from(_decrypt(receiverRawSK, data));
+export function decrypt(
+  receiverRawSK: string | Uint8Array,
+  data: Uint8Array
+): Uint8Array {
+  const decrypted = _decrypt(receiverRawSK, data);
+  return IS_BUFFER_SUPPORTED ? Buffer.from(decrypted) : decrypted;
 }
 
 function _decrypt(