+++ src/keys/PublicKey.ts
@@ -1,6 +1,8 @@
 import { bytesToHex, equalBytes } from "@noble/ciphers/utils";
 
 import type { EllipticCurve } from "../config.js";
+
+import { IS_BUFFER_SUPPORTED } from "../types.js";
 import { convertPublicKeyFormat, getSharedKey, hexToPublicKey } from "../utils/index.js";
 import type { PrivateKey } from "./PrivateKey.js";
 
@@ -17,13 +19,15 @@ export class PublicKey {
   }
 
   /** @deprecated - use `PublicKey.toBytes(false)` instead. You may also need `Buffer.from`. */
-  get uncompressed(): Buffer {
-    return Buffer.from(this._uncompressed); // TODO: delete
+  get uncompressed(): Uint8Array {
+    // TODO: delete
+    return IS_BUFFER_SUPPORTED ? Buffer.from(this._uncompressed) : this._uncompressed;
   }
 
   /** @deprecated - use `PublicKey.toBytes()` instead. You may also need `Buffer.from`. */
-  get compressed(): Buffer {
-    return Buffer.from(this.data); // TODO: delete
+  get compressed(): Uint8Array {
+    // TODO: delete
+    return IS_BUFFER_SUPPORTED ? Buffer.from(this.data) : this.data;
   }
 
   constructor(data: Uint8Array, curve?: EllipticCurve) {