+++ CHANGELOG.md
@@ -5,7 +5,7 @@
 
 - Bump dependencies
 - Prepare for v0.5.0
-  - API change: Use `Buffer` only when it's available
+  - API change: Return a Node.js `Buffer` only when available (otherwise a `Uint8Array`)
 
 ## 0.4.17